+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+
+## React Native Kommunicate Chat v2.0.1
+- Added support for listening to Events - iOS
+```javascript
+    const eventEmitter = new NativeEventEmitter(RNKommunicateChat);
+    const eventListener = eventEmitter.addListener("onMessageReceived", event => {
+        console.log("onMessageReceived" );
+    });
+ ```
+ - Exposed CloseConversationScreen() to close the conversation screen
 ## React Native Kommunicate Chat v2.0.0
 - Added support for listening to Events - Android only
  ```javascript